test: decompose PROJECT readiness proof and prune retired route journeys
