+++ backend/scripts/run_test_lanes.py
@@ -243,6 +243,8 @@ LANES = (
             "tests/projects/guide_compilation/test_repository_attempts.py",
             "tests/projects/guide_compilation/test_repository_persistence.py",
             "tests/projects/test_locked_policy_context.py",
+            "tests/projects/test_activation_readiness.py",
+            "tests/projects/test_retired_submission_derivation_route.py",
             "tests/test_projects.py",
         ),
     ),
